+++ packages/app/hivecommand-native/src/views/setup/context.tsx
@@ -1,13 +1,45 @@
 import React from 'react'
 
+export type StateUpdateFn = (stateUpdate: (state: GlobalState) => any) => void;
+
+// export type StateUpdate = (key: string, value: any) => void;
+
+export interface GlobalState {
+    subscriptionMap?: {
+        path: string,
+        tag: string
+    }[],
+    controlLayout?: {
+        interface: {
+            id: string,
+            nodes: any[],
+            edges: any[]
+        }
+        devices: {
+            id: string;
+            tag: string;
+            type: {
+                tagPrefix?: string;
+                state: {
+                    key: string
+                }[];
+            }
+        }[]
+    },
+    networkLayout?: any[]
+    deviceMap?: {
+        path: string,
+        tag: string
+    }[]
+}
 export const SetupContext = React.createContext<{
     state?: any;
     setState?: any;
-    globalState?: any;
-    setGlobalState?: any;
+    globalState?: GlobalState
+    setGlobalState?: StateUpdateFn;
 }>({
 
-    
+
 });
 
 export const SetupProvider = SetupContext.Provider;