+++ packages/communication/iot-client/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.15-alpha.129](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.128...v1.4.15-alpha.129) (2022-12-06)
+
+**Note:** Version bump only for package @hive-command/iot-client
+
+
+
+
+
+## [1.4.15-alpha.128](https://github.com/TheTechCompany/HiveCommand/compare/v0.1.3...v1.4.15-alpha.128) (2022-12-05)
+
+**Note:** Version bump only for package @hive-command/iot-client
+
+
+
+
+
+## [1.4.15-alpha.127](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.126...v1.4.15-alpha.127) (2022-12-04)
+
+**Note:** Version bump only for package @hive-command/iot-client
+
+
+
+
+
 ## [1.4.15-alpha.126](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.125...v1.4.15-alpha.126) (2022-12-04)
 
 **Note:** Version bump only for package @hive-command/iot-client